+++ two_d_game_ai/entities/bot.py
@@ -43,7 +43,6 @@ class Bot(GenericEntity):
     def __init__(self, world: World, name: str, pos: tuple[float, float]) -> None:
         super().__init__(world, name, pos)
         self._velocity_v = Vector2(0, 0)
-        self._destination = tuple[float, float] | None
         self._destination_v: Vector2 | None = None
         self.heading: Bearing = Bearing(Bot.INITIAL_HEADING_DEGREES)
         self.known_bots: set[Bot] = set()
@@ -63,7 +62,7 @@ class Bot(GenericEntity):
     @destination.setter
     def destination(self, value: tuple[float, float]) -> None:
         self.stop()
-        self._destination_v = Vector2(value)
+        self.destination_v = Vector2(value)
 
     @property
     def destination_v(self) -> Vector2 | None:
@@ -72,8 +71,9 @@ class Bot(GenericEntity):
 
     @destination_v.setter
     def destination_v(self, value: Vector2) -> None:
-        self.stop()
-        self._destination_v = value
+        if value is not None and self.world.point_is_inside_world_bounds(value):
+            self.stop()
+            self._destination_v = value
 
     @property
     def is_at_destination(self) -> bool: